requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2==2.9.3
pytest-django==4.5.2
pytest==6.2.5
black==22.6.0
